--- .env-template
@@ -1,2 +1,0 @@
-FLASK_APP=app
-FLASK_DEBUG=True
