+++ templates/base.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>CriptoSimulator</title>
+    <link rel="stylesheet" href="/static/lib/bootstrap.min.css" />
+    <script src="/static/lib/bootstrap.bundle.min.js"></script>
+    <link rel="stylesheet" href="/static/css/styles.css" />
+  </head>
+
+  <body>
+    {% include 'components/header.html' %} {% include 'components/navbar.html'
+    %}
+
+    <main>{% block content %} {% endblock content %}</main>
+
+    {% include 'components/footer.html' %}
+  </body>
+</html>
