+++ templates/components/footer.html
@@ -1,0 +1,1 @@
+<footer class="pie-pagina">2023 CriptoSimulator</footer>
