+++ templates/components/header.html
@@ -1,0 +1,1 @@
+<header>All transactions</header>
