+++ templates/components/navbar.html
@@ -1,0 +1,5 @@
+<nav>
+  <a href="/">Home</a>
+  |
+  <a href="/purchase">Purchase</a>
+</nav>
