--- transaction/views/index.html
@@ -1,44 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Document</title>
-  </head>
-  <body>
-    <h1>Transactions</h1>
-
-    <section class="section-transactions">
-      <button id="recharge-button">Reload</button>
-      <table class="paper">
-        <thead>
-          <tr>
-            <th class="table-title">ID</th>
-            <th class="table-title">Date</th>
-            <th class="table-title">Time</th>
-            <th class="table-title">From</th>
-            <th class="table-title">Quantity</th>
-            <th class="table-title">To</th>
-            <th class="table-title">From Quantity</th>
-            <th class="table-title">Purchased Units</th>
-          </tr>
-        </thead>
-        <tbody id="cuerpo-tabla">
-          {% for transaction in transactions %}
-          <tr>
-            <td>{{transaction.id}}</td>
-            <td>{{transaction.time}}</td>
-            <td>{{transaction.date}}</td>
-            <td>{{transaction._from}}</td>
-            <td>{{transaction.qty}}</td>
-            <td>{{transaction._to}}</td>
-            <td>{{transaction.from_qty}}</td>
-            <td>{{transaction.pu}}</td>
-            {% endfor %}
-          </tr>
-        </tbody>
-      </table>
-    </section>
-  </body>
-</html>
